Extract base and Pairlock from models file
