+++ freqtrade/persistence/base.py
@@ -1,0 +1,7 @@
+
+from typing import Any
+
+from sqlalchemy.orm import declarative_base
+
+
+_DECL_BASE: Any = declarative_base()
